move enum to gta related enums file
