+++ src/backend/looped/vehicle/seatbelt.cpp
@@ -1,4 +1,5 @@
 #include "backend/looped_command.hpp"
+#include "gta/enums.hpp"
 #include "natives.hpp"
 
 namespace big
@@ -10,13 +11,13 @@ namespace big
 		virtual void on_tick() override
 		{
 			PED::SET_PED_CONFIG_FLAG(self::ped, 32, false);
-			PED::SET_PED_CAN_BE_KNOCKED_OFF_VEHICLE(self::ped, eKnockOffVehicle::KNOCKOFFVEHICLE_NEVER);
+			PED::SET_PED_CAN_BE_KNOCKED_OFF_VEHICLE(self::ped, (int)eKnockOffVehicle::KNOCKOFFVEHICLE_NEVER);
 		}
 
 		virtual void on_disable() override
 		{
 			PED::SET_PED_CONFIG_FLAG(self::ped, 32, true);
-			PED::SET_PED_CAN_BE_KNOCKED_OFF_VEHICLE(self::ped, eKnockOffVehicle::KNOCKOFFVEHICLE_DEFAULT);
+			PED::SET_PED_CAN_BE_KNOCKED_OFF_VEHICLE(self::ped, (int)eKnockOffVehicle::KNOCKOFFVEHICLE_DEFAULT);
 		}
 	};
