+++ src/core/enums.hpp
@@ -143,14 +143,6 @@ namespace big
 		UNK2                 = 1 << 16
 	};
 
-	enum class eKnockOffVehicle
-	{
-		KNOCKOFFVEHICLE_DEFAULT,
-		KNOCKOFFVEHICLE_NEVER,
-		KNOCKOFFVEHICLE_EASY,
-		KNOCKOFFVEHICLE_HARD
-	};
-
 	enum class ePedTask
 	{
 		TASK_NONE,
